Исправлены ошибки. Теперь работают методы удаления данных

+++ Program.cs
@@ -1,5 +1,7 @@
 ﻿using ConsoleAppEntityFramework.Models;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace ConsoleAppEntityFramework
 {
@@ -12,7 +14,7 @@ namespace ConsoleAppEntityFramework
             context = new AppDBContext();
         }
         #region InputData
-        public static void InputData(Customer customer)
+        private static void InputData(Customer customer)
         {
             Console.Write("Input name: ");
             customer.Name = Console.ReadLine();
@@ -25,7 +27,6 @@ namespace ConsoleAppEntityFramework
             
             context.Customers.Add(customer);
             context.SaveChanges();
-            context.Dispose();
         }
         private static void InputData(Payment payment)
         {
@@ -38,7 +39,6 @@ namespace ConsoleAppEntityFramework
 
             context.Payments.Add(payment);
             context.SaveChanges();
-            context.Dispose();
         }
         private static void InputData(Product product)
         {
@@ -53,7 +53,6 @@ namespace ConsoleAppEntityFramework
 
             context.Products.Add(product);
             context.SaveChanges();
-            context.Dispose();
         }
         private static void InputData(Log log)
         {
@@ -64,12 +63,11 @@ namespace ConsoleAppEntityFramework
             
             context.Logs.Add(log);
             context.SaveChanges();
-            context.Dispose();
         }
         #endregion
 
-        #region RemoveData
-        private static void RemoveData(Customer customer)
+        #region _RemoveData
+        private static void _RemoveData(Customer customer)
         {
             Console.Write("Input name: ");
             customer.Name = Console.ReadLine();
@@ -80,11 +78,18 @@ namespace ConsoleAppEntityFramework
             Console.Write("Input email: ");
             customer.Email = Console.ReadLine();
 
-            context.Customers.Remove(customer);
-            context.SaveChanges();
-            context.Dispose();
-        }   
-        private static void RemoveData(Payment payment)
+            var custom1 = context.Customers.FirstOrDefault(cust => cust.Name == customer.Name);
+
+            if (custom1.Name == customer.Name)
+            {
+                context.Entry(custom1).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
+                context.Customers.Remove(customer);
+                context.SaveChanges();
+            }
+            else Console.WriteLine("Таких данных нет в базе");
+
+        }
+        private static void _RemoveData(Payment payment)
         {
             Console.Write("Input ChatID: ");
             payment.ChatID = Console.ReadLine();
@@ -93,11 +98,17 @@ namespace ConsoleAppEntityFramework
             Console.Write("Input Sum: ");
             payment.Sum = Int32.Parse(Console.ReadLine());
 
-            context.Payments.Remove(payment);
-            context.SaveChanges();
-            context.Dispose();
+            var paym1 = context.Payments.FirstOrDefault(p => p.Sum == payment.Sum);
+
+            if (paym1.Sum == payment.Sum)
+            {
+                context.Entry(paym1).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
+                context.Payments.Remove(payment);
+                context.SaveChanges();
+            }
+            else Console.WriteLine("Таких данных нет в базе");
         }
-        private static void RemoveData(Product product)
+        private static void _RemoveData(Product product)
         {
             Console.Write("Input name: ");
             product.Name = Console.ReadLine();
@@ -107,26 +118,39 @@ namespace ConsoleAppEntityFramework
             product.Price = Int32.Parse(Console.ReadLine());
             Console.Write("Input Discription: ");
             product.Discription = Console.ReadLine();
-            
-            context.Products.Remove(product);
-            context.SaveChanges();
-            context.Dispose();
+
+            var prod1 = context.Products.FirstOrDefault(products => products.Name == product.Name);
+
+            if (prod1.Name == product.Name)
+            {
+                context.Entry(prod1).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
+                context.Products.Remove(prod1);
+                context.SaveChanges();
+            }
+            else Console.WriteLine("Таких данных нет в базе");
         }
-        private static void RemoveData(Log log)
+        private static void _RemoveData(Log log)
         {
             Console.Write("Input ChatID: ");
             log.ChatID = Console.ReadLine();
             Console.Write("Input command: ");
             log.Command = Console.ReadLine();
- 
-            context.Logs.Remove(log);
-            context.SaveChanges();
-            context.Dispose();
+
+            var log1 = context.Logs.FirstOrDefault(logs => logs.Command == log.Command);
+
+            if (log1.Command == log.Command)
+            {
+                context.Entry(log1).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
+                context.Logs.Remove(log);
+                context.SaveChanges();
+            }
+            else Console.WriteLine("Таких данных нет в базе");
         }
         #endregion
         static void Main(string[] args)
         {
-            
+             string ChooseNumber = null;
+             string ChooseNumber2 = null;
             while (true)
             {
                 Console.WriteLine("Введите номер команды");
@@ -134,10 +158,10 @@ namespace ConsoleAppEntityFramework
                     "1. Product" + "\n"
                     + "2. Customer" + "\n"
                     + "3. Log" + "\n"
-                    + "4. Payment");
+                    + "4. Payment" + "\n"
+                    + "5. Выход из программы");
 
-                string ChooseNumber = Console.ReadLine();
-                string ChooseNumber2;
+                ChooseNumber = Console.ReadLine();
 
                 switch (ChooseNumber)
                 {
@@ -151,7 +175,7 @@ namespace ConsoleAppEntityFramework
                                 InputData(product);
                                 break;
                             case "2":
-                                RemoveData(product);
+                                _RemoveData(product);
                                 break;
                         }
                         break;
@@ -165,7 +189,7 @@ namespace ConsoleAppEntityFramework
                                 InputData(customer);
                                 break;
                             case "2":
-                                RemoveData(customer);
+                                _RemoveData(customer);
                                 break;
                         }
                         break;
@@ -180,7 +204,7 @@ namespace ConsoleAppEntityFramework
                                 InputData(log);
                                 break;
                             case "2":
-                                RemoveData(log);
+                                _RemoveData(log);
                                 break;
                         }
                         break;
@@ -194,11 +218,13 @@ namespace ConsoleAppEntityFramework
                                 InputData(payment);
                                 break;
                             case "2":
-                                RemoveData(payment);
+                                _RemoveData(payment);
                                 break;
                         }
                         break;
-                    default: System.Environment.Exit(0);
+                    case "5":
+                        context.Dispose();
+                        System.Environment.Exit(0);
                         break;
                 }
             }
